Refactor basic pubsub functionality
