+++ src/index.js
@@ -4,20 +4,20 @@
 
 'use strict'
 
-const Pubsub = require('libp2p-pubsub')
+const assert = require('assert')
 const pull = require('pull-stream')
 const lp = require('pull-length-prefixed')
 const nextTick = require('async/nextTick')
 const { utils } = require('libp2p-pubsub')
 
+const BasicPubsub = require('./pubsub')
 const { MessageCache } = require('./messageCache')
-const assert = require('assert')
 
 const { rpc } = require('./message')
 const constants = require('./constants')
 const errcode = require('err-code')
 
-class GossipSub extends Pubsub {
+class GossipSub extends BasicPubsub {
   /**
    * @param {Object} libp2p
    * @constructor
@@ -65,11 +65,6 @@ class GossipSub extends Pubsub {
      *
      */
     this.messageCache = new MessageCache(constants.GossipSubHistoryGossip, constants.GossipSubHistoryLength)
-
-    /**
-     * A set of subscriptions
-     */
-    this.subscriptions = new Set()
   }
 
   /**
@@ -101,97 +96,14 @@ class GossipSub extends Pubsub {
   }
 
   /**
-   * When a peer has dialed into another peer, it sends its subscriptions to it.
-   * @override
-   * @param {PeerInfo} peerInfo
-   * @param {Connection} conn
-   * @param {Function} callback
+   * Handles an rpc control message from a peer
    *
-   * @returns {void}
-   *
-   */
-  _onDial (peerInfo, conn, callback) {
-    super._onDial(peerInfo, conn, (err) => {
-      if (err) return callback(err)
-      const idB58Str = peerInfo.id.toB58String()
-      const peer = this.peers.get(idB58Str)
-      if (peer && peer.isWritable) {
-        // Immediately send my own subscription to the newly established conn
-        peer.sendSubscriptions(this.subscriptions)
-      }
-      nextTick(() => callback())
-    })
-  }
-
-  /**
-   * Processes a peer's connection to another peer.
-   *
-   * @param {String} idB58Str
-   * @param {Connection} conn
    * @param {Peer} peer
-   *
-   * @returns {void}
-   *
-   */
-  _processConnection (idB58Str, conn, peer) {
-    pull(
-      conn,
-      lp.decode(),
-      pull.map((data) => rpc.RPC.decode(data)),
-      pull.drain(
-        (rpc) => this._onRpc(idB58Str, rpc),
-        (err) => this._onConnectionEnd(idB58Str, peer, err)
-      )
-    )
-  }
-
-  /**
-   * Handles an rpc request from a peer
-   *
-   * @param {String} idB58Str
-   * @param {Object} rpc
+   * @param {rpc.RPC} rpc
    * @returns {void}
    */
-  _onRpc (idB58Str, rpc) {
-    if (!rpc) {
-      return
-    }
-
-    const peer = this.peers.get(idB58Str)
-    if (!peer) {
-      return
-    }
-
-    this.log('rpc from', idB58Str)
+  _handleRpcControl (peer, rpc) {
     const controlMsg = rpc.control
-    const subs = rpc.subscriptions
-    const msgs = rpc.msgs
-
-    if (subs.length) {
-      // update peer subscriptions
-      peer.updateSubscriptions(subs)
-      subs.forEach((subOptMsg) => {
-        const t = subOptMsg.topicID
-
-        if (!this.topics.has(t)) {
-          this.topics.set(t, new Set())
-        }
-
-        const topicSet = this.topics.get(t)
-        if (subOptMsg.subscribe) {
-          // subscribe peer to new topic
-          topicSet.add(peer)
-        } else {
-          // unsubscribe from existing topic
-          topicSet.delete(peer)
-        }
-      })
-      this.emit('meshsub:subscription-change', peer.info, peer.topics, subs)
-    }
-
-    if (msgs.length) {
-      this._processRpcMessages(utils.normalizeInRpcMessages(msgs))
-    }
 
     if (!controlMsg) {
       return
@@ -211,85 +123,38 @@ class GossipSub extends Pubsub {
   }
 
   /**
-   * Process incoming messages,
+   * Process incoming message,
    * emitting locally and forwarding on to relevant floodsub and gossipsub peers
    */
-  _processRpcMessages (msgs) {
-    msgs.forEach((msg) => {
-      const seqno = utils.msgId(msg.from, msg.seqno)
-      // Have we seen this message before> if so, ignore
-      if (this.seenCache.has(seqno)) {
+  _processRpcMessage (msg) {
+    super._processRpcMessage(msg)
+    const topics = msg.topicIDs
+
+    // Emit to floodsub peers
+    this.peers.forEach((peer) => {
+      if (peer.info.protocols.has(constants.FloodSubID) &&
+        peer.info.id.toB58String() !== msg.from &&
+        utils.anyMatch(peer.topics, topics) &&
+        peer.isWritable
+      ) {
+        peer.sendMessages(utils.normalizeOutRpcMessages([msg]))
+        this.log('publish msg on topics - floodsub', topics, peer.info.id.toB58String())
+      }
+    })
+
+    // Emit to peers in the mesh
+    topics.forEach((topic) => {
+      if (!this.mesh.has(topic)) {
         return
       }
-
-      this.seenCache.put(seqno)
-
-      const topics = msg.topicIDs
-
-      // Emit to self
-      this._emitMessages(topics, [msg])
-
-      // Emit to floodsub peers
-      this.peers.forEach((peer) => {
-        if (peer.info.protocols.has(constants.FloodSubID) &&
-          peer.info.id.toB58String() !== msg.from &&
-          utils.anyMatch(peer.topics, topics) &&
-          peer.isWritable
-        ) {
-          peer.sendMessages(utils.normalizeOutRpcMessages([msg]))
-          this.log('publish msg on topics - floodsub', topics, peer.info.id.toB58String())
-        }
-      })
-
-      // Emit to peers in the mesh
-      topics.forEach((topic) => {
-        if (!this.mesh.has(topic)) {
+      this.mesh.get(topic).forEach((peer) => {
+        if (!peer.isWritable || peer.info.id.toB58String() === msg.from) {
           return
         }
-        this.mesh.get(topic).forEach((peer) => {
-          if (!peer.isWritable || peer.info.id.toB58String() === msg.from) {
-            return
-          }
-          peer.sendMessages(utils.normalizeOutRpcMessages([msg]))
-          this.log('publish msg on topic - meshsub', topic, peer.info.id.toB58String())
-        })
+        peer.sendMessages(utils.normalizeOutRpcMessages([msg]))
+        this.log('publish msg on topic - meshsub', topic, peer.info.id.toB58String())
       })
     })
-  }
-
-  _emitMessages (topics, messages) {
-    topics.forEach((topic) => {
-      if (this.subscriptions.has(topic)) {
-        messages.forEach((message) => {
-          this.emit(topic, message)
-        })
-      }
-    })
-  }
-
-  /**
-   * Returns a buffer of a RPC message that contains a control message
-   *
-   * @param {Array<rpc.RPC.Message>} msgs
-   * @param {Array<rpc.RPC.ControlIHave>} ihave
-   * @param {Array<rpc.RPC.ControlIWant>} iwant
-   * @param {Array<rpc.RPC.ControlGraft>} graft
-   * @param {Array<rpc.RPC.Prune>} prune
-   *
-   * @returns {rpc.RPC}
-   *
-   */
-  _rpcWithControl (msgs, ihave, iwant, graft, prune) {
-    return {
-      subscriptions: [],
-      msgs: msgs || [],
-      control: {
-        ihave: ihave || [],
-        iwant: iwant || [],
-        graft: graft || [],
-        prune: prune || []
-      }
-    }
   }
 
   /**
@@ -481,7 +346,6 @@ class GossipSub extends Pubsub {
       this.lastpub = new Map()
       this.gossip = new Map()
       this.control = new Map()
-      this.subscriptions = new Set()
       this._heartbeatTimer.cancel(() => {
         this._heartbeatTimer = null
         callback()
@@ -490,31 +354,17 @@ class GossipSub extends Pubsub {
   }
 
   /**
-   * Subscribes to topics
+   * Join topics
    * @param {Array<string>|string} topics
    * @returns {void}
    */
-  subscribe (topics) {
+  join (topics) {
     assert(this.started, 'GossipSub has not started')
-
     topics = utils.ensureArray(topics)
 
-    const newTopics = topics.filter((topic) => !this.subscriptions.has(topic))
-    if (newTopics.length === 0) {
-      return
-    }
+    this.log('JOIN %s', topics)
 
-    this.log('JOIN %s', newTopics)
-
-    // Broadcast SUBSCRIBE to all peers
-    this.peers.forEach((peer) => {
-      peer.sendSubscriptions(newTopics)
-    })
-
-    newTopics.forEach((topic) => {
-      // set subscription
-      this.subscriptions.add(topic)
-
+    topics.forEach((topic) => {
       // Send GRAFT to mesh peers
       const fanoutPeers = this.fanout.get(topic)
       if (fanoutPeers) {
@@ -533,29 +383,17 @@ class GossipSub extends Pubsub {
   }
 
   /**
-   * Leaves a topic
+   * Leave topics
    *
    * @param {Array<string>|string} topics
    * @returns {void}
    */
-  unsubscribe (topics) {
+  leave (topics) {
     topics = utils.ensureArray(topics)
 
-    const unTopics = topics.filter((topic) => this.subscriptions.has(topic))
-    if (!unTopics.length) {
-      return
-    }
     this.log('LEAVE %s', topics)
 
-    // Broadcast UNSUBSCRIBE to all peers
-    this.peers.forEach((peer) => {
-      peer.sendUnsubscriptions(topics)
-    })
-
-    unTopics.forEach((topic) => {
-      // delete subscription
-      this.subscriptions.delete(topic)
-
+    topics.forEach((topic) => {
       // Send PRUNE to mesh peers
       const meshPeers = this.mesh.get(topic)
       if (meshPeers) {
@@ -568,38 +406,8 @@ class GossipSub extends Pubsub {
     })
   }
 
-  /**
-   * Publishes messages to all subscribed peers
-   *
-   * @param {Array<string>|string} topics
-   * @param {Array<any>|any} messages
-   * @returns {void}
-   */
-  publish (topics, messages) {
-    assert(this.started, 'GossipSub has not started')
-    this.log('publish', topics, messages)
-
-    topics = utils.ensureArray(topics)
-    messages = utils.ensureArray(messages)
-
-    const from = this.libp2p.peerInfo.id.toB58String()
-
-    const buildMessage = (msg) => {
-      const seqno = utils.randomSeqno()
-      const msgObj = {
-        from: from,
-        data: msg,
-        seqno: seqno,
-        topicIDs: topics
-      }
-      this.messageCache.put(msgObj)
-      this.seenCache.put(msgObj.seqno)
-      return msgObj
-    }
-
-    const msgObjs = utils.normalizeOutRpcMessages(messages.map(buildMessage))
-
-    msgObjs.forEach((msgObj) => {
+  _publish (messages) {
+    messages.forEach((msgObj) => {
       // @type Set<string>
       const tosend = new Set()
       msgObj.topicIDs.forEach((topic) => {
@@ -902,63 +710,6 @@ class GossipSub extends Pubsub {
     let gossip = this.gossip.get(peer)
     gossip = gossip.concat(controlIHaveMsgs)
     this.gossip.set(peer, gossip)
-  }
-
-  /**
-   * Given a topic, returns up to count peers subscribed to that topic
-   *
-   * @param {String} topic
-   * @param {Number} count
-   * @returns {Set<Peer>}
-   *
-   */
-  _getPeers (topic, count) {
-    const peersInTopic = this.topics.get(topic)
-    if (!peersInTopic) {
-      return new Set()
-    }
-
-    // Adds all peers using GossipSub protocol
-    let peers = []
-    peersInTopic.forEach((peer) => {
-      if (peer.info.protocols.has(constants.GossipSubID)) {
-        peers.push(peer)
-      }
-    })
-
-    // Pseudo-randomly shuffles peers
-    peers = this._shufflePeers(peers)
-    if (count > 0 && peers.length > count) {
-      peers = peers.slice(0, count)
-    }
-
-    peers = new Set(peers)
-    return peers
-  }
-
-  /**
-   * Pseudo-randomly shuffles peers
-   *
-   * @param {Array<Peers>} peers
-   * @returns {Array<Peers>}
-   */
-  _shufflePeers (peers) {
-    if (peers.length <= 1) {
-      return peers
-    }
-
-    for (let i = 0; i < peers.length; i++) {
-      const randInt = () => {
-        return Math.floor(Math.random() * Math.floor(peers.length))
-      }
-
-      const j = randInt()
-      const tmp = peers[i]
-      peers[i] = peers[j]
-      peers[j] = tmp
-
-      return peers
-    }
   }
 
   /**
